Create working REST APIs for Products
This includes creating Entities for Products, Product Types and Product Brands.
Generate migration and update the database.
Seed the database with sample data from local Json files.
Implementing a repository pattern to access and use the database using the context.
Updating the controller to create new endpoints for Products, Brands and Types.